move spammer to TimedSpammer
